api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure.ai.formrecognizer
requests
jsons
azure-identity
azure-storage-blob
azure-core
openai==0.27.2
azure-kusto-data==4.0.2
pandas
pyodbc
azure-search-documents==11.4.0b3
python-dotenv
numpy==1.24.3